+++ raven/librarian/app.py
@@ -61,6 +61,7 @@ with timer() as tim:
     from ..common.audio import player as audio_player
     from ..common.audio import recorder as audio_recorder
     from ..common import bgtask
+    from ..common import docextract
     from ..common import utils as common_utils
 
     from ..common.gui import animation as gui_animation
@@ -244,10 +245,12 @@ with timer() as tim:
     retriever, scanner = hybridir.setup(docs_dir=docs_dir,
                                         recursive=librarian_config.llm_docs_dir_recursive,
                                         db_dir=db_dir,
+                                        exts=librarian_config.llm_docs_exts,
+                                        callback=docextract.extract_text,  # extract plaintext from PDFs (and read text files)
                                         embedding_model_name=librarian_config.qa_embedding_model,
                                         local_model_loader_fallback=False)  # Librarian requires Raven-server for other reasons, too
 
-    logger.info(f"RAG document store is at '{str(librarian_config.llm_docs_dir)}' (put your plain-text documents here).")
+    logger.info(f"RAG document store is at '{str(librarian_config.llm_docs_dir)}' (put your text or PDF documents here).")
     # The retriever's `documents` attribute must be locked before accessing.
     with retriever.datastore_lock:
         plural_s = "s" if len(retriever.documents) != 1 else ""